update logout, profile for admin page, validate for admin

+++ modules/users_management/listUser.php
@@ -230,9 +230,14 @@ layoutAdminUseInclude("header", $dataTitle);
                                         echo '<td>';
                                         echo '<a class="btn btn-warning mx-2" title="Edit user information" href="?module=users_management&action=editUser&user_id=' . $row['user_id'] . '"><i class="fa-solid fa-pen-to-square" style="color: #ffffff;"></i></a>';
                                         echo '</td>';
-                                        echo '<td>';
-                                        echo '<a class="btn btn-danger" title="Delete user" href="?module=users_management&action=deleteUser&user_id=' . $row['user_id'] . '"><i class="fa-solid fa-trash"></i></a>';
-                                        echo '</td>';
+                                        if ($user_id == $row['user_id']) {
+                                            echo '<td><span class="badge bg-danger">You</span></td>';
+                                        } else {
+                                            echo '<td>';
+                                            echo '<a class="btn btn-danger" title="Delete user" href="?module=users_management&action=deleteUser&user_id=' . $row['user_id'] . '"><i class="fa-solid fa-trash"></i></a>';
+                                            echo '</td>';
+                                        }
+
                                         echo '</tr>';
                                     }
                                 }
